feat: Add API Tools Management Page and Functionality
- Created a new ToolsPage component to manage API tools.
- Implemented APIToolsTable component for displaying and managing external tools.
- Added functionality for creating, editing, viewing, and deleting tools with appropriate modals.
- Integrated API calls for fetching tools and handling errors with fallback to mock data.
- Updated menu items to remove the "Doanh thu" option.
